requirements.txt
apache-airflow
mysql-connector-python
couchdb